test(app): move feishu config tests into dedicated module
